+++ astropy/conftest.py
@@ -4,7 +4,9 @@ This file contains pytest configuration settings that are astropy-specific
 (i.e.  those that would not necessarily be shared by affiliated packages
 making use of astropy's test runner).
 """
+import os
 import builtins
+import tempfile
 from importlib.util import find_spec
 
 import astropy
@@ -44,6 +46,20 @@ def pytest_configure(config):
         matplotlibrc_cache.update(matplotlib.rcParams)
         matplotlib.rcdefaults()
 
+    # Make sure we use temporary directories for the config and cache
+    # so that the tests are insensitive to local configuration. Note that this
+    # is also set in the test runner, but we need to also set it here for
+    # things to work properly in parallel mode
+
+    builtins._xdg_config_home_orig = os.environ.get('XDG_CONFIG_HOME')
+    builtins._xdg_cache_home_orig = os.environ.get('XDG_CACHE_HOME')
+
+    os.environ['XDG_CONFIG_HOME'] = tempfile.mkdtemp('astropy_config')
+    os.environ['XDG_CACHE_HOME'] = tempfile.mkdtemp('astropy_cache')
+
+    os.mkdir(os.path.join(os.environ['XDG_CONFIG_HOME'], 'astropy'))
+    os.mkdir(os.path.join(os.environ['XDG_CACHE_HOME'], 'astropy'))
+
 
 def pytest_unconfigure(config):
     builtins._pytest_running = False
@@ -51,6 +67,16 @@ def pytest_unconfigure(config):
     if HAS_MATPLOTLIB:
         matplotlib.rcParams.update(matplotlibrc_cache)
         matplotlibrc_cache.clear()
+
+    if builtins._xdg_config_home_orig is None:
+        os.environ.pop('XDG_CONFIG_HOME')
+    else:
+        os.environ['XDG_CONFIG_HOME'] = builtins._xdg_config_home_orig
+
+    if builtins._xdg_cache_home_orig is None:
+        os.environ.pop('XDG_CACHE_HOME')
+    else:
+        os.environ['XDG_CACHE_HOME'] = builtins._xdg_cache_home_orig
 
 
 PYTEST_HEADER_MODULES['Cython'] = 'cython'