+++ docs/conftest.py
@@ -1,14 +1,12 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
-# This file is the main file used when running tests with pytest directly,
-# in particular if running e.g. ``pytest docs/``.
+# This file needs to be included here to make sure commands such
+# as ``python setup.py test ... -t docs/...`` works, since this
+# will ignore the conftest.py file at the root of the repository
+# and the one in astropy/conftest.py
 
 import os
 import tempfile
-
-pytest_plugins = [
-    'astropy.tests.plugins.display',
-]
 
 # Make sure we use temporary directories for the config and cache
 # so that the tests are insensitive to local configuration.